Create books with Rails forms & JSON
'Create' starts with a Rails form and passes input from the form to
Angular as a JSON object.  Once the new JSON object is saved it
invokes the getBooks function to load the list of books.

The question is:  if we're loading the list of books without the
getBooks() function in the previous commit - isn't there a way to do so
with loading the books after a book is created?  Perhaps it's because in
the previous commit there was no interaction between Rails & Angular but
in saving books you need both - hence the call to load the books....